+++ src/Model/Config/StandardParametersData.php
@@ -32,24 +32,32 @@ class StandardParametersData
     private $indentsCommentsWithoutParent;
 
     /**
+     * @var string[]
+     */
+    private $alphabeticalPrioritizedKeys;
+
+    /**
      * @param int $depth
      * @param int $indents
      * @param int $level
      * @param string $serviceAliasingType
      * @param string $indentsCommentsWithoutParent
+     * @param string[] $alphabeticalPrioritizedKeys
      */
     public function __construct(
         int $depth,
         int $indents,
         int $level,
         string $serviceAliasingType,
-        string $indentsCommentsWithoutParent
+        string $indentsCommentsWithoutParent,
+        array $alphabeticalPrioritizedKeys
     ) {
         $this->depth = $depth;
         $this->indents = $indents;
         $this->level = $level;
         $this->serviceAliasingType = $serviceAliasingType;
         $this->indentsCommentsWithoutParent = $indentsCommentsWithoutParent;
+        $this->alphabeticalPrioritizedKeys = $alphabeticalPrioritizedKeys;
     }
 
     /**
@@ -90,5 +98,13 @@ class StandardParametersData
     public function getIndentsCommentsWithoutParent(): string
     {
         return $this->indentsCommentsWithoutParent;
+    }
+
+    /**
+     * @return string[]
+     */
+    public function getAlphabeticalPrioritizedKeys(): array
+    {
+        return $this->alphabeticalPrioritizedKeys;
     }
 }