+++ tests/Model/YamlServiceAliasing/YamlServiceAliasingFixerTest.php
@@ -72,6 +72,6 @@ class YamlServiceAliasingFixerTest extends TestCase
      */
     private function getStandardsParametersData(string $type): StandardParametersData
     {
-        return new StandardParametersData(4, 4, 2, $type, YamlStandardConfigDefinition::CONFIG_PARAMETERS_INDENTS_COMMENTS_WITHOUT_PARENT_VALUE_DEFAULT);
+        return new StandardParametersData(4, 4, 2, $type, YamlStandardConfigDefinition::CONFIG_PARAMETERS_INDENTS_COMMENTS_WITHOUT_PARENT_VALUE_DEFAULT, []);
     }
 }